requirements.txt
# Data
kaggle==1.6.17
pandas==2.2.3
numpy==1.26.4

# NLP
nltk==3.9.1
spacy==3.7.5
gensim==4.3.3
vaderSentiment==3.3.2
transformers==4.44.2
torch==2.4.1
scikit-learn==1.5.2

# Visualization
plotly==5.24.1
streamlit==1.39.0
matplotlib==3.9.2
seaborn==0.13.2
wordcloud==1.9.3

# Utilities
tqdm==4.66.5
python-dotenv==1.0.1
pyarrow==17.0.0
